requirements.txt.txt
streamlit
supabase>=2,<3
python-dotenv
pandas
numpy
